fix: string comparison in pwd() fixed
The for loop in pwd() could never find the PWD environment variable as it was doing an incorrect string comparison : the whole environment variable was being compared to string "PWD".
To fix this, I had to implement a function that returns the substring of a given string.

Now pwd is fully functional.